Change scheduler in remove-background/run.py
I was interested in multi-threading, so I installed this branch of cellbender.
Running it with multi-threading didn't work due to an AttributeError, as OneCycleLR
is not in pyro.optim (I guess they moved OneCycleLR in pytorch). I fixed it to the best of my knowledge,
but I'm new to these modules and this process in general, so any feedback is appreciated.

remove-background works now with multi-threading, but it has a warning and an error:
- a warning from torch: UserWarning: Detected call of `lr_scheduler.step()` before `optimizer.step()`. In PyTorch 1.1.0 and later, you should call them in the opposite order: `optimizer.step()` before `lr_scheduler.step()`.  Failure to do this will result in PyTorch skipping the first value of the learning rate schedule. See more details at https://pytorch.org/docs/stable/optim.html#how-to-adjust-learning-rate
- an error from torch after the 75th epoch: ValueError: Tried to step 7951 times. The specified number of total steps is 7950

This is my first time contributing to an open source project, so any feedback is appreciated!