Add multiprocessing option for the MCKP estimator (#201)
* Improve cell prob plot and remove features not analyzed

This was very confusing and led to what looked like a horrible bug: disagreement between the HTML report plot and the PDF plot.  But it's because the report was plotting all features, not just features analyzed.  The report now also plots features analyzed

* Refactor sparse utils and eliminate duplication

* Add multiprocessing MCKP estimation as command line option

On Mac laptop this is not faster in small tests